Gas gates
