+++ app/App/Panel/Local/style/index.styl
@@ -130,9 +130,29 @@
 .viewLicense:hover
   color outerspace08
 
+.discordInvite:hover
+  background ghostF
+
+.discordInvite
+  margin 32px 8px 0px 8px 
+  font-size 12px
+  font-weight 400
+  height 32px
+  border-radius 16px
+  display flex
+  justify-content center
+  align-items center
+  width calc(100% - 16px)
+  background lowlight
+  color outerspace08
+
+  .discordLink
+    font-weight 600
+    padding-left 5px
+
 .quitFrame
   padding 10px
-  margin-top 15px
+  margin-top 22px
 
   .quitFrameButton
     width 100%
