+++ app/App/Panel/Main/Signer/Settings/style/index.styl
@@ -111,6 +111,14 @@
     padding 5px 0px 0px 0px
     z-index 20
 
+  .keyCommand
+    margin-left 2px
+    padding 0px 4px 2px 4px
+    font-weight 400
+    border-radius 4px
+    font-size 12px
+    border 1px solid outerspace02
+
   .nodeProviderStatus
     display flex
     justify-content center
