+++ app/App/Panel/style/index.styl
@@ -177,11 +177,9 @@
     align-items center
     font-weight 500
     font-size 11px
-    background highlighter
     height 16px
     border-radius 8px
-    transform translate(-50px, 11px)
-    box-shadow cardDrop, 0px 8px 1px -6px thick
+    transform translate(-50px, 12px)
 
     .panelMenuDataDivide
       position absolute
