+++ package.json
@@ -102,9 +102,9 @@
     "tar": "6.0.5",
     "trezor-connect": "8.1.19",
     "usb": "1.6.3",
-    "uuid": "8.3.1",
+    "uuid": "8.3.2",
     "web3-utils": "1.3.0",
-    "ws": "7.4.0",
+    "ws": "7.4.1",
     "zxcvbn": "4.4.2"
   },
   "devDependencies": {
@@ -120,12 +120,12 @@
     "@babel/preset-env": "7.12.7",
     "@babel/preset-react": "7.12.7",
     "babel-eslint": "10.1.0",
-    "electron": "11.0.3",
+    "electron": "11.0.4",
     "electron-builder": "22.9.1",
     "electron-notarize": "1.0.0",
     "ganache-cli": "6.12.1",
     "jest": "26.6.3",
-    "node-watch": "0.7.0",
+    "node-watch": "0.7.1",
     "parcel-bundler": "1.12.4",
     "standard": "16.0.3",
     "stylus": "0.54.8"
